bot/requirements.txt
aiogram==3.7.*
SQLAlchemy>=2.0
asyncpg>=0.29
python-dotenv>=1.0
